pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-experiment-scaffold"
version = "0.1.0"
description = "Production ML experiment scaffold for single-GPU homelab (RTX 5070 Ti, Blackwell)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "Tremayne Timms"}]
keywords = ["ml", "experiment", "fine-tuning", "grpo", "scaffold"]

dependencies = [
    "PyYAML>=6.0,<7.0",
    "numpy>=1.24.0,<3.0",
]

[project.optional-dependencies]
llm = [
    "torch>=2.10.0",
    "transformers>=5.0.0,<6.0",
    "trl>=0.29.0,<1.0",
    "peft>=0.10.0,<1.0",
    "accelerate>=0.33.0,<2.0",
    "datasets>=2.18.0,<5.0",
    "bitsandbytes>=0.44.0,<1.0",
    "wandb>=0.18.0,<1.0",
    "lm_eval>=0.4.0,<1.0",
    "sentencepiece>=0.2.0,<1.0",
]
vision = [
    "torch>=2.10.0",
    "torchvision>=0.20.0",
    "timm>=1.0.0,<2.0",
    "wandb>=0.18.0,<1.0",
]
dev = [
    "pytest>=7.0,<9.0",
    "pytest-cov>=4.0,<6.0",
    "ruff>=0.4.0,<1.0",
    "hypothesis>=6.100.0,<7.0",
    "pre-commit>=3.7.0,<5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.ruff]
target-version = "py311"
line-length = 100
exclude = ["checkpoints", "results", "wandb"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"src/train.py" = ["I001"]
"src/models.py" = ["I001"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
omit = ["src/train.py", "src/models.py"]

[tool.coverage.report]
fail_under = 70
